Verify requesting all member
