+++ backend/interface/database_api/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod api;
 
 mod connection;
-mod database;
 mod frb_generated;
 mod json_field_conversion;
 mod schema;
