+++ backend/interface/database_api/src/test_database_common.rs
@@ -1,3 +1,8 @@
+use std::fs;
+
+use backend_testing::testing;
+use sqlx::{Database, Pool, any::install_default_drivers, pool::PoolConnection};
+
 use crate::connection::OrmConnection;
 
 trait GetCurrentDBName {
@@ -67,4 +72,26 @@ impl IntoDieselConnection for sqlx::pool::PoolConnection<sqlx::MySql> {
             diesel::MysqlConnection::establish(&db_url).expect("Could not establish connection"),
         )
     }
+}
+
+pub async fn setup_test<DB>(sqlx_pool: Pool<DB>) -> OrmConnection
+where
+    DB: Database,
+    PoolConnection<DB>: IntoDieselConnection,
+{
+    testing::setup_test();
+
+    install_default_drivers(); // FIXME Required?
+
+    let sqlx_connection = sqlx_pool.acquire().await.unwrap();
+    let mut diesel_connection = sqlx_connection.into_diesel_connection().await;
+
+    let fixture_sql_content = fs::read_to_string("src/fixtures/allsupportedtypes.sql").unwrap();
+    diesel_connection.execute_sql(fixture_sql_content);
+
+    diesel_connection
+}
+
+pub fn tear_down(expected_num_severe_messages: usize) {
+    testing::tear_down(expected_num_severe_messages);
 }
